Minor cleanup.
Broke supporting class out into own source file. No structural code
changes; no spec changes.

RSpec: 713 examples, 0 failures, 1 pending
RuboCop: 230 files inspected, no offenses detected
